requirements.txt
beautifulsoup4==4.13.4
boto3==1.39.3
botocore==1.39.3
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cssselect==1.3.0
cycler==0.12.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
fonttools==4.58.5
idna==3.10
jieba3k==0.35.1
jmespath==1.0.1
joblib==1.5.1
kiwisolver==1.4.8
lightgbm==4.6.0
lxml==6.0.0
lxml_html_clean==0.4.2
matplotlib==3.10.3
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.3.1
packaging==25.0
pandas==2.3.0
pillow==11.3.0
pyarrow==20.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-file==2.1.0
s3transfer==0.13.0
scikit-learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.7
threadpoolctl==3.6.0
tinysegmenter==0.3
tldextract==5.3.0
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
